remove hardware/ibom/ibom.html
The updated schematics, Gerbers, etc. are kept in
the "hardware/Project Outputs for ov_ftdi/" and it is
confusing if there is another "BOM/Bill of Materials-ov_ftdi.csv".

Users can always go back in commit history to obtain the older
versions, but master should not contain renderings of old versions.