Report untruncated packet size to host
Do not immediately proceed to writing header when packet larger than
MAX_PACKET_SIZE is captured on the bus. Keep counting packet length so
the value can be reported to the host.

When Truncated flag is set the size reported in header is untruncated
packet size and only MAX_PACKET_SIZE bytes are included. In order to not
increase header size, and thus limit the maximum capture throughput, the
MAX_PACKET_SIZE value must match between gateware and host.

Because change modifies gateware to host software protocol and hence is
backwards incompatible. I believe the gateware should be considered
integral part of software package and therefore I do not consider this
protocol incompatibility to be a problem, especially that the gateware
is only loaded to volatile memory.

Reporting actual length as seen on the bus allows USBInterpreter to
display for example "Truncated 2 bytes" instead of "Unexpected ERR CRC".
Similarly, when exporting to pcap format, Wireshark can display "[Packet
size limited during capture]" instead of reporting CRC errors.